Chapter12/virtualenvs/myproject_env/project/django-myproject/requirements/base.txt
Pillow~=5.2.0
mysqlclient~=1.3.0
Django~=2.1.0
django-crispy-forms~=1.7.0
xhtml2pdf~=0.2.3
django-haystack~=2.8.0
whoosh~=2.7.0
sorl-thumbnail~=12.4.0
openpyxl~=2.5.0
django-watermark~=0.1.8
python-dotenv~=0.8.0
social-auth-app-django~=2.1.0
python-memcached~=1.59.0
django-redis~=4.9.0
django-cms~=3.6.0
## if Django CMS 3.6 has not yet been released, use the development version:
#-e git://github.com/divio/django-cms.git@develop#egg=django-cms
django-mptt~=0.9.1
django-mptt-admin~=0.6.0
django-treebeard~=4.3.0
requests~=2.19.1
django-tastypie~=0.14.2
djangorestframework~=3.8.2
transliterate~=1.10.2
# For Python < 3.3:
#mock~=2.0.0
cookiecutter~=1.6.0
